[BE] Long 데이터 타입 변경

+++ spring-data-elasticsearch/src/main/java/com/example/elasticsearch/controller/ReviewController.java
@@ -26,20 +26,20 @@ public class ReviewController {
 
     /* 후기 리스트 조회 */
     @GetMapping("{restaurantSeq}")
-    public ResponseEntity<?> getReviewList(@PathVariable Integer restaurantSeq) {
+    public ResponseEntity<?> getReviewList(@PathVariable Long restaurantSeq) {
         return ResponseEntity.ok(reviewService.getReviewList(restaurantSeq));
     }
 
 
     /* 후기 상세 조회 */
     @GetMapping("detail/{reviewSeq}")
-    public ResponseEntity<?> getReview(@PathVariable Integer reviewSeq) {
+    public ResponseEntity<?> getReview(@PathVariable Long reviewSeq) {
         return ResponseEntity.ok(reviewService.getReview(reviewSeq));
     }
 
     /* 후기 수정 */
     @PutMapping("detail/{reviewSeq}")
-    public ResponseEntity<?> modifyReview(@RequestBody ModifyReviewDto modifyReviewDto, @PathVariable Integer reviewSeq) {
+    public ResponseEntity<?> modifyReview(@RequestBody ModifyReviewDto modifyReviewDto, @PathVariable Long reviewSeq) {
         return ResponseEntity.ok(reviewService.modifyReview(modifyReviewDto));
     }
